Move zulip API to a different mod
